requirements.txt
Flask==2.3.3
mysql-connector-python==9.3.0
pandas==2.2.1
numpy==1.26.2
scikit-learn==1.3.2
statsmodels==0.14.4
openpyxl==3.1.2
python-dotenv==1.0.0
cryptography==41.0.7
patsy==1.0.1